added ExternalLink and ReceiptLink to renamed helpers/links.tsx file in routes. Updated twotone topbar to be more consistent with the color of the border above the white section. this was messed up due to the sitewhite box-sizing change and forced sizing.
