+++ packages/augur-simplified/src/modules/common/top-nav.styles.less
@@ -17,14 +17,15 @@
   > section {
     align-items: center;
     box-sizing: border-box;
-    height: @size-80;
-    min-height: @size-80;
-    max-height: @size-80;
+    height: @size-78;
+    min-height: @size-78;
+    max-height: @size-78;
     display: flex;
     flex-flow: row nowrap;
     flex: 1;
     grid-gap: @size-16;
     width: 100%;
+    margin: auto 0;
 
     &:first-of-type {
       max-width: @size-912;
